Refactor HTTP and WS applications
