+++ blunt.cabal
@@ -22,16 +22,15 @@ source-repository head
 library
     exposed-modules:
         Blunt
-        Blunt.Actions
         Blunt.Application
+        Blunt.HTTP
         Blunt.Main
         Blunt.Markup
-        Blunt.Middleware
         Blunt.Pointfree
         Blunt.Pointful
-        Blunt.Router
         Blunt.Script
         Blunt.Style
+        Blunt.WebSocket
     build-depends:
         base ==4.*,
         aeson ==0.8.*,
